add export stubs in a `cjs-module-lexer`-friendly way
This approach is more resilient to changes in React, because we're only adding the exports we want to stub, meaning that this doesn't need to get updated every time react adds a new API -- we only need to touch this when a new *client-only* API is added. In addition, adding fake `module.exports.foo` assignments makes this `cjs-module-lexer`-friendly, so we get nice named exports for these (while also preserving any exports coming from React).